refactor(telegram): split outbound send operations
